+++ cw_5/cw_5/DTOs/Requests/EnrollStudentPromotions.cs
@@ -1,8 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations;
-using System.Linq;
-using System.Threading.Tasks;
+﻿using System.ComponentModel.DataAnnotations;
 
 namespace cw_5.DTOs.Requests
 {
