+++ cw_5/cw_5/DTOs/Responses/EnrollStudentResponse.cs
@@ -1,8 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
+﻿
 namespace cw_5.DTOs.Responses
 {
     public class EnrollStudentResponse
