+++ cw_5/cw_5/Services/IStudentDbService.cs
@@ -1,14 +1,11 @@
 ﻿using cw_5.DTOs.Requests;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
+using Microsoft.AspNetCore.Mvc;
 
 namespace cw_5.Services
 {
     public interface IStudentDbService
     {
-        void EnrollStudents(EnrollStudentRequests request);
-        void EnrollStudentPromotions(EnrollStudentPromotions request);
+        IActionResult EnrollStudents(EnrollStudentRequests newStudent);
+        IActionResult EnrollStudentPromotions(EnrollStudentPromotions promotion);
     }
 }
